requirements.txt
accelerate==0.19.0
anthropic==0.2.9
google-generativeai==0.1.0
grpcio-status==1.33.2
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
protobuf==3.19.6
torch==1.13.1
transformers==4.29.1
chess==1.9.4
scikit-learn==1.2.2
matplotlib
pandas==2.0.3
